* commit '1bc507a4738d64855c7389ba4ec68f308201e615':
  JDWP: check thread state on event suspension

+++ jdwp/src/test/java/org/apache/harmony/jpda/tests/jdwp/share/JDWPProxyTestCase.java
@@ -61,7 +61,7 @@ public abstract class JDWPProxyTestCase extends JDWPSyncTestCase {
         // Set breakpoint.
         String debuggeeSignature = getDebuggeeClassSignature();
         long classId = getClassIDBySignature(debuggeeSignature);
-        long requestId = debuggeeWrapper.vmMirror.setBreakpointAtMethodBegin(
+        int requestId = debuggeeWrapper.vmMirror.setBreakpointAtMethodBegin(
                 classId, "breakpointMethodFromProxy");
 
         // Signal debuggee to continue execution.
